Add new attributes to storage record
Namely, the Spotify track ID and the context in which the song was
played.  The track ID I'm putting in here in case I want to do some
future analysis on this data (because finding the track based on
artist name/album name/track name would work, but is another API call
and could result in ambiguous results).  The context I'm putting in here
because if I decide to do more with this in the future, I might want
to downplay stuff I listened to through Discover Weekly or something.